pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verbal-direction"
version = "0.1.0"
description = "Voice-controlled Claude Code session manager"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Sam French"},
]
keywords = ["claude", "voice", "tts", "stt", "ai", "terminal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "claude-agent-sdk>=0.1.40",
    "faster-whisper>=1.0.0",
    "piper-tts>=1.4.0",
    "silero-vad>=5.1",
    "sounddevice>=0.4.6",
    "ollama>=0.4.0",
    "textual>=0.50.0",
    "click>=8.1.0",
    "numpy>=1.24.0",
]

[project.scripts]
vd = "verbal_direction.__main__:cli"

[project.urls]
Homepage = "https://github.com/samuelfrench/verbal-direction"
Repository = "https://github.com/samuelfrench/verbal-direction"

[tool.setuptools.packages.find]
include = ["verbal_direction*"]
